requirements.txt
mcp>=1.0.0
firebase-admin>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
